% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCrawls.R
\name{compareCrawls}
\alias{compareCrawls}
\title{Compare crawl over a crawl - NOT IMPLEMENTED}
\usage{
compareCrawls(crawlId)
}
\arguments{
\item{:}{crawlID}
}
\value{
Json
}
\description{
Compare crawl over a crawl - NOT IMPLEMENTED
}
\details{
ResCode
401 : invalid key or ip
429 : Too Many Requests
500 : Internal error
550 : Error while generating guide.
250 : Guide generation on the way. Please wait a few minutes.
}
\examples{
\dontrun{
crawls <- compareCrawls(crawlID)
}
}
\author{
Vincent Terrasi
}
