% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataviz.R
\name{export_formattableWidget}
\alias{export_formattableWidget}
\title{export_formattableWidget(out,"metrics.png", width=500)}
\usage{
export_formattableWidget(w, file, width = 400, background = "white",
  delay = 0.2)
}
\description{
export_formattableWidget(out,"metrics.png", width=500)
}
