% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCrawl.R
\name{getCrawl}
\alias{getCrawl}
\title{Get a crawl}
\usage{
getCrawl(crawlId)
}
\arguments{
\item{crawlId}{Id of your crawl}
}
\value{
Json
}
\description{
Get a crawl
}
\details{
<http://developer.oncrawl.com/#Get-a-crawl>

ResCode
400 : Returned when the request has incompatible values or does not match the API specification.
401 : Returned when the request is not authenticated.
403 : Returned the current quota does not allow the action to be performed.
404 : Returned when any of resources referred in the request is not found.
403 : Returned when the request is authenticated but the action is not allowed.
409 : Returned when the requested operation is not allowed for current state of the resource.
500 : Internal error

The HTTP response is JSON object with a single crawl key containing the crawl’s data
}
\examples{
\dontrun{
initAPI()
project <- getCrawl("YOURCRAWLID")
}

}
\author{
Vincent Terrasi
}
