% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProject.R
\name{getProject}
\alias{getProject}
\title{Get a project}
\usage{
getProject(projectId)
}
\arguments{
\item{projectId}{Id of your project}
}
\value{
Json
}
\description{
Get a project
}
\details{
<http://developer.oncrawl.com/#Get-a-project>

ResCode
400 : Returned when the request has incompatible values or does not match the API specification.
401 : Returned when the request is not authenticated.
403 : Returned the current quota does not allow the action to be performed.
404 : Returned when any of resource(s) referred in the request is not found.
403 : Returned when the request is authenticated but the action is not allowed.
409 : Returned when the requested operation is not allowed for current state of the resource.
500 : Internal error

The HTTP response is JSON object with three keys:
- A project key with the project’s data
- A crawl_configs key with a list of all the project’s crawl configurations.
- A crawls key with a list of all project’s crawl.
}
\examples{
\dontrun{
initAPI()
project <- getProject(YOURPROJECTID)
}
}
\author{
Vincent Terrasi
}
