% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initAPI.R
\name{initAPI}
\alias{initAPI}
\title{Prepare Token for API calls}
\usage{
initAPI()
}
\value{
character if no error with API authentification
}
\description{
Prepare Token for API calls
}
\examples{
\donttest{
initAPI()
}

}
\author{
Vincent Terrasi
}
