% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listLogs.R
\name{listLogs}
\alias{listLogs}
\title{List all pages from logs monitoring}
\usage{
listLogs(projectId)
}
\arguments{
\item{:}{projectID}
}
\value{
Json
}
\description{
List all pages from logs monitoring
}
\details{
ResCode
400 : Returned when the request has incompatible values or does not match the API specification.
401 : Returned when the request is not authenticated.
403 : Returned the current quota does not allow the action to be performed.
404 : Returned when any of resource(s) referred in the request is not found.
403 : Returned when the request is authenticated but the action is not allowed.
409 : Returned when the requested operation is not allowed for current state of the resource.
500 : Internal error
}
\examples{
\dontrun{
pages <- listLogs(projectID)
}
}
\author{
Vincent Terrasi
}
