% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPages.R
\name{listPages}
\alias{listPages}
\title{List all pages from a crawl}
\usage{
listPages(crawlId)
}
\arguments{
\item{crawlId}{ID of your crawl}
}
\value{
Json
}
\description{
List all pages from a crawl
}
\details{
<http://developer.oncrawl.com/#Data-types>

ResCode
400 : Returned when the request has incompatible values or does not match the API specification.
401 : Returned when the request is not authenticated.
403 : Returned the current quota does not allow the action to be performed.
404 : Returned when any of resource(s) referred in the request is not found.
403 : Returned when the request is authenticated but the action is not allowed.
409 : Returned when the requested operation is not allowed for current state of the resource.
500 : Internal error
}
\examples{
\dontrun{
pages <- listPages(YOURCRAWLID)
}

}
\author{
Vincent Terrasi
}
