% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPages.R
\name{listPagesAggs}
\alias{listPagesAggs}
\title{Get Aggregate Queries for a specific OQL}
\usage{
listPagesAggs(crawlId, oqlList)
}
\arguments{
\item{crawlId}{ID of your crawl}

\item{oqlList}{json of your OQL}
}
\value{
Json
}
\description{
Get Aggregate Queries for a specific OQL
}
\details{
<http://developer.oncrawl.com/#OnCrawl-Query-Language>

ResCode
400 : Returned when the request has incompatible values or does not match the API specification.
401 : Returned when the request is not authenticated.
403 : Returned the current quota does not allow the action to be performed.
404 : Returned when any of resource(s) referred in the request is not found.
403 : Returned when the request is authenticated but the action is not allowed.
409 : Returned when the requested operation is not allowed for current state of the resource.
500 : Internal error
}
\examples{
\dontrun{
agg <- listPagesAggs(YOURCRAWLID, YOURJSON)
page_crawled = agg[[1]]$rows
}

}
\author{
Vincent Terrasi
}
