% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listProjects.R
\name{listProjects}
\alias{listProjects}
\title{List all projects}
\usage{
listProjects(limit = 100)
}
\arguments{
\item{limit}{number of projects}
}
\value{
Json
}
\description{
List all projects
}
\details{
<http://developer.oncrawl.com/#List-projects>

ResCode
400 : Returned when the request has incompatible values or does not match the API specification.
401 : Returned when the request is not authenticated.
403 : Returned the current quota does not allow the action to be performed.
404 : Returned when any of resources referred in the request is not found.
403 : Returned when the request is authenticated but the action is not allowed.
409 : Returned when the requested operation is not allowed for current state of the resource.
500 : Internal error
}
\examples{
\donttest{
initAPI()
projects <- listProjects()
}

}
\author{
Vincent Terrasi
}
