% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataviz.R
\name{oncrawlCreateDashboard}
\alias{oncrawlCreateDashboard}
\title{Create a dashboard}
\usage{
oncrawlCreateDashboard(dataset, namefile, width, pathfile = tempdir())
}
\arguments{
\item{dataset}{Dataset with 3 columns : date, unique name of metric, value}

\item{namefile}{Filename for the export}

\item{width}{Width of your picture}

\item{pathfile}{string. Optional. If not specified, the intermediate files are created under \code{TEMPDIR}, with the assumption that directory is granted for written permission.}
}
\value{
a graph
}
\description{
Create a dashboard
}
\examples{
\dontrun{
oncrawlCreateDashboard(res, "metric.png", 500, ".")
}

}
\author{
Vincent Terrasi
}
