% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataviz.R
\name{oncrawlCreateGraph}
\alias{oncrawlCreateGraph}
\title{Create a graph}
\usage{
oncrawlCreateGraph(dataset, namefile, width, height,
  pathfile = tempdir())
}
\arguments{
\item{dataset}{dataset generated by DALEX package}

\item{namefile}{the filename for the export}

\item{width}{width of your picture}

\item{height}{height of your picture}

\item{pathfile}{string. Optional. If not specified, the intermediate files are created under \code{TEMPDIR}, with the assumption that directory is granted for written permission.}
}
\value{
file
}
\description{
Create a graph
}
\examples{
\dontrun{
oncrawlCreateGraph(res, "metric.png", width=5, height=4, ".")
}

}
\author{
Vincent Terrasi
}
