% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oncrawlCreateSegmentation}
\alias{oncrawlCreateSegmentation}
\title{Cluster URLs}
\usage{
oncrawlCreateSegmentation(list_urls, namefile)
}
\arguments{
\item{list_urls}{your urls}

\item{namefile}{the filename for the CSV export}
}
\value{
data.frame
}
\description{
Cluster URLs
}
\examples{
\dontrun{
cluster(DF)
}
}
\author{
Vincent Terrasi
}
