% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oncrawlExplainModel}
\alias{oncrawlExplainModel}
\title{Explain XGBoost Model by plotting each importance variables}
\usage{
oncrawlExplainModel(model, x, y, max = 10)
}
\arguments{
\item{model}{your XgBoost model}

\item{x}{your training data}

\item{y}{your predicted data}

\item{max}{the number of importance variable you want to explain}
}
\value{
graphs
}
\description{
Explain XGBoost Model by plotting each importance variables
}
\examples{
\dontrun{
oncrawlExplainModel(DF)
}
}
\author{
Vincent Terrasi
}
