% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oncrawlSplitURL}
\alias{oncrawlSplitURL}
\title{Split URLs}
\usage{
oncrawlSplitURL(list_urls, limit = 15)
}
\arguments{
\item{list_urls}{your urls}

\item{limit}{the maximum of URLS you want}
}
\value{
data.frame
}
\description{
Split URLs
}
\examples{
\dontrun{
oncrawlSplitURL(DF)
}
}
\author{
Vincent Terrasi
}
