% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataviz.R
\name{export_formattableWidget}
\alias{export_formattableWidget}
\title{Transform HTML widget into picture}
\usage{
export_formattableWidget(w, file, width = 400, background = "white",
  delay = 0.2)
}
\arguments{
\item{w}{HTML content to print}

\item{file}{A vector of names of output files. Should end with .png, .pdf, or .jpeg. If several screenshots have to be taken and only one filename is provided, then the function appends the index number of the screenshot to the file name.}

\item{width}{Viewport width. This is the width of the browser "window".}

\item{background}{Background color for web page}

\item{delay}{Time to wait before taking screenshot, in seconds. Sometimes a longer delay is needed for all assets to display properly.}
}
\description{
Transform HTML widget into picture
}
