% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initAPI.R
\name{initAPI}
\alias{initAPI}
\title{Prepare Token for API calls}
\usage{
initAPI(path)
}
\arguments{
\item{path}{path of your conf file}
}
\value{
ok if no error with API authentification
}
\description{
Prepare Token for API calls
}
\details{
Example file for oncrawl_configuration.txt

key = 5516LP29W5Q9XXXXXXXXXXXXOEUGWHM9
debug = FALSE
api = https://app.oncrawl.com/api/v2/
}
\examples{
\dontrun{
initAPI("oncrawl_configuration.txt")
}

}
\author{
Vincent Terrasi
}
