% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-model.R
\name{oncrawlExplainModel}
\alias{oncrawlExplainModel}
\title{Explain XGBoost Model by displaying each importance variables}
\usage{
oncrawlExplainModel(model, x, y, max = 10, path = tempdir())
}
\arguments{
\item{model}{your XgBoost model}

\item{x}{your training data}

\item{y}{your predicted data}

\item{max}{the number of importance variable you want to explain}

\item{path}{path of your conf file}
}
\value{
graphs
}
\description{
Explain XGBoost Model by displaying each importance variables
}
\examples{
\dontrun{
list <- oncrawlTrainModel(dataset,200)
oncrawlExplainModel(list$model, list$x, list$y, 3)
}
}
\author{
Vincent Terrasi
}
