% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-model.R
\name{oncrawlTrainModel}
\alias{oncrawlTrainModel}
\title{Train XGBoost Model}
\usage{
oncrawlTrainModel(dataset, nround = 300, verbose = 1)
}
\arguments{
\item{dataset}{your data frame}

\item{nround}{number of iterations}

\item{verbose}{display errors ?}
}
\value{
a list with your ML model, your training data
}
\description{
Train XGBoost Model
}
\examples{
\dontrun{
list <- oncrawlTrainModel(dataset)
plot(list$roc)
print(list$matrix)
}

}
\author{
Vincent Terrasi
}
